A2Y-146 remove anort lint option in :common module
