+++ domain/src/main/java/com/yapp/itemfinder/domain/model/Category.kt
@@ -1,0 +1,9 @@
+package com.yapp.itemfinder.domain.model
+
+data class Category(
+    var name: String = "default_name",
+    override var type: CellType = CellType.CATEGORY_CELL
+) : Data() {
+
+    fun goCategoryPage() = handler.invoke(this)
+}
