+++ domain/src/main/java/com/yapp/itemfinder/domain/model/Data.kt
@@ -10,12 +10,5 @@ open class Data(
 ) : Parcelable {
 
     var handler: DataHandler = { }
-}
 
-data class Category(
-    var name: String = "default_name",
-    override var type: CellType = CellType.CATEGORY_CELL
-) : Data() {
-
-    fun goCategoryPage() = handler.invoke(this)
 }
